Wrap logic in function and module
Move all of `site_listing.py` to within `get_sites()`, which takes a
state postal code (eg MD) as a parameter and returns a list of dicts
describing the USGS stream gages in that state.

Add `__init__.py` to make the folder a python module and allow it to
be imported with `from stream_gages.site_listing import get_sites`.

Add `.gitignore` copied from the github template.